refactor: use new numerators
