+++ src/components/config-tabs/numerators/Numerators.js
