+++ src/components/config-tabs/numerators/index.js
@@ -1,1 +1,1 @@
-export { Numerators } from './NewNumerators.js'
+export { Numerators } from './Numerators.js'
